Use Postcode.toOutcode rather than custom postcodeArea method
